+++ pkg/mads/reconcile/reconciler.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	envoy_core "github.com/envoyproxy/go-control-plane/envoy/api/v2/core"
-	envoy_cache "github.com/envoyproxy/go-control-plane/pkg/cache"
+	envoy_cache "github.com/envoyproxy/go-control-plane/pkg/cache/v2"
 
 	util_xds "github.com/Kong/kuma/pkg/util/xds"
 )