+++ pkg/mads/server/mads.go
@@ -19,8 +19,8 @@ import (
 
 	envoy "github.com/envoyproxy/go-control-plane/envoy/api/v2"
 	envoy_core "github.com/envoyproxy/go-control-plane/envoy/api/v2/core"
-	envoy_cache "github.com/envoyproxy/go-control-plane/pkg/cache"
-	envoy_server "github.com/envoyproxy/go-control-plane/pkg/server"
+	envoy_cache "github.com/envoyproxy/go-control-plane/pkg/cache/v2"
+	envoy_server "github.com/envoyproxy/go-control-plane/pkg/server/v2"
 )
 
 type Server interface {