+++ pkg/util/cache/cache.go
@@ -6,12 +6,12 @@ import (
 	"github.com/golang/protobuf/ptypes"
 
 	v2 "github.com/envoyproxy/go-control-plane/envoy/api/v2"
-	ctl_cache "github.com/envoyproxy/go-control-plane/pkg/cache"
+	ctl_cache "github.com/envoyproxy/go-control-plane/pkg/cache/v2"
 )
 
 func ToDeltaDiscoveryResponse(s ctl_cache.Snapshot) (*v2.DeltaDiscoveryResponse, error) {
 	resp := &v2.DeltaDiscoveryResponse{}
-	for _, rs := range []ctl_cache.Resources{s.Endpoints, s.Clusters, s.Routes, s.Listeners, s.Secrets} {
+	for _, rs := range s.Resources {
 		for _, name := range sortedResourceNames(rs) {
 			r := rs.Items[name]
 			pbany, err := ptypes.MarshalAny(r)