+++ pkg/util/xds/callbacks_chain_test.go
@@ -10,7 +10,7 @@ import (
 	. "github.com/Kong/kuma/pkg/util/xds"
 
 	envoy "github.com/envoyproxy/go-control-plane/envoy/api/v2"
-	envoy_xds "github.com/envoyproxy/go-control-plane/pkg/server"
+	envoy_xds "github.com/envoyproxy/go-control-plane/pkg/server/v2"
 )
 
 var _ = Describe("CallbacksChain", func() {