+++ pkg/util/xds/watchdog_callbacks.go
@@ -6,7 +6,7 @@ import (
 
 	envoy "github.com/envoyproxy/go-control-plane/envoy/api/v2"
 	envoy_core "github.com/envoyproxy/go-control-plane/envoy/api/v2/core"
-	envoy_xds "github.com/envoyproxy/go-control-plane/pkg/server"
+	envoy_xds "github.com/envoyproxy/go-control-plane/pkg/server/v2"
 
 	util_watchdog "github.com/Kong/kuma/pkg/util/watchdog"
 )